refactor: keep update reconciliation within size budgets
